#20: Consolidated the StreamingRecordBuilder and macro to be consistent with the other record builders.
